bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gnews = "^0.3.7"
pandas = "^2.2.2"
requests = "^2.32.3"
pypdf = "^4.3.1"
langchain-community = "^0.2.11"
ipykernel = "^6.29.5"
langchain-openai = "^0.1.20"
langchain-chroma = "^0.1.2"
asyncstdlib = "^3.12.4"
python-dotenv = "^1.0.1"
pymongo = {extras = ["srv"], version = "^4.8.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
